+++ misc.js
@@ -4,20 +4,19 @@ const glob = require('glob');
  * Returns all files matching the wild card pattern
  * @param {*} filenameWildcard filename wildcard pattern
  */
-const getMatchingfiles = (filenameWildcard) => {
-    return new Promise((resolve, reject) => {
-        glob(filenameWildcard, (err, files) => {
-            if (err) {
-                console.log(`Failed getting filenames, error = ${err}`);
+const getMatchingFiles = (filenameWildcard) => new Promise((resolve, reject) => {
+    glob(filenameWildcard, (err, files) => {
+        if (err) {
+            // eslint-disable-next-line no-console
+            console.error(`Failed getting filenames, error = ${err}`);
 
-                reject(err);
-            } else {
-                resolve(files);
-            }
-        });
-    })
-};
+            reject(err);
+        } else {
+            resolve(files);
+        }
+    });
+});
 
 module.exports = {
-    getMatchingfiles
+    getMatchingFiles
 };