+++ test/002.testGraph.js
@@ -4,8 +4,8 @@ const { JGFContainer } = require('../jgfContainer');
 describe('Graph', () => {
     describe('#add graph node', () => {
         it('should add a simple node to a graph', () => {
-            const container = new JGFContainer(singleGraph = true);
-            const graph = container.graph;
+            const container = new JGFContainer(true);
+            const { graph } = container;
 
             const nodeId = 'lebron-james#2254';
             const nodeLabel = 'LeBron James';
@@ -14,11 +14,11 @@ describe('Graph', () => {
             assert.equal(1, graph.nodes.length);
             assert.equal(nodeId, graph.nodes[0].id);
             assert.equal(nodeLabel, graph.nodes[0].label);
-        })
+        });
 
         it('should add a node to a graph, with meta data', () => {
-            const container = new JGFContainer(singleGraph = true);
-            const graph = container.graph;
+            const container = new JGFContainer(true);
+            const { graph } = container;
 
             const nodeId = 'kevin-durant#4497';
             const nodeLabel = 'Kevin Durant';
@@ -33,12 +33,12 @@ describe('Graph', () => {
             assert.equal(1, graph.nodes.length);
             assert.equal('Power Forward', graph.nodes[0].metadata.position);
             assert.equal(35, graph.nodes[0].metadata.shirt);
-        })
+        });
 
 
         it('should throw an exception when adding a node that already exists', () => {
-            const container = new JGFContainer(singleGraph = true);
-            const graph = container.graph;
+            const container = new JGFContainer(true);
+            const { graph } = container;
 
             const nodeId = 'kevin-durant#4497';
             const nodeLabel = 'Kevin Durant';
@@ -46,11 +46,11 @@ describe('Graph', () => {
             graph.addNode(nodeId, nodeLabel);
 
             assert.throw(() => graph.addNode(nodeId, nodeLabel), Error, 'A node already exists');
-        })
+        });
 
         it('should throw an exception when adding nodes that already exist', () => {
-            const container = new JGFContainer(singleGraph = true);
-            const graph = container.graph;
+            const container = new JGFContainer(true);
+            const { graph } = container;
 
             const nodeId = 'kevin-durant#4497';
             const nodeLabel = 'Kevin Durant';
@@ -69,14 +69,13 @@ describe('Graph', () => {
             ];
 
             assert.throw(() => graph.addNodes(moreNodes), Error, 'A node already exists');
-        })
-
-    })
+        });
+    });
 
     describe('#update graph node', () => {
         it('should update a node\'s label', () => {
-            const container = new JGFContainer(singleGraph = true);
-            const graph = container.graph;
+            const container = new JGFContainer(true);
+            const { graph } = container;
 
             const nodeId = 'kevin-love#0000';
             const nodeLabel = 'Kevin Lofe';
@@ -85,14 +84,13 @@ describe('Graph', () => {
             const correctLabel = 'Kevin Love';
             graph.updateNode(nodeId, correctLabel);
             assert.equal(correctLabel, graph.nodes[0].label);
-        })
-
-    })
+        });
+    });
 
     describe('#removeNode', () => {
         it('should remove a node', () => {
-            const container = new JGFContainer(singleGraph = true);
-            const graph = container.graph;
+            const container = new JGFContainer(true);
+            const { graph } = container;
 
             const nodeId = 'kevin-durant#4497';
             const nodeLabel = 'Kevin Durant';
@@ -101,20 +99,20 @@ describe('Graph', () => {
 
             graph.removeNode(nodeId);
             assert.equal(0, graph.nodes.length, 'After removeNode there should be zero nodes');
-        })
+        });
 
         it('should throw an exception when removing a non existant node', () => {
-            const container = new JGFContainer(singleGraph = true);
-            const graph = container.graph;
+            const container = new JGFContainer(true);
+            const { graph } = container;
 
             assert.throws(() => graph.removeNode('some dummy id'), 'A node doesn\'t exist');
-        })
-    })
+        });
+    });
 
     describe('#getNode', () => {
         it('should lookup a node by id', () => {
-            const container = new JGFContainer(singleGraph = true);
-            const graph = container.graph;
+            const container = new JGFContainer(true);
+            const { graph } = container;
 
             const nodeId = 'kevin-durant#4497';
             const nodeLabel = 'Kevin Durant';
@@ -124,20 +122,20 @@ describe('Graph', () => {
             const node = graph.getNode(nodeId);
             assert(node !== null);
             assert(node.id === nodeId);
-        })
+        });
 
         it('should throw an exception when looking up a non existant node', () => {
-            const container = new JGFContainer(singleGraph = true);
-            const graph = container.graph;
+            const container = new JGFContainer(true);
+            const { graph } = container;
 
             assert.throws(() => graph.getNode('some dummy id'), 'A node doesn\'t exist');
-        })
-    })
+        });
+    });
 
     describe('#addGraphEdge', () => {
         it('should add a simple edge to a graph', () => {
-            const container = new JGFContainer(singleGraph = true);
-            const graph = container.graph;
+            const container = new JGFContainer(true);
+            const { graph } = container;
 
             const node1Id = 'lebron-james#2254';
             const node1Label = 'LeBron James';
@@ -156,13 +154,13 @@ describe('Graph', () => {
 
             assert.equal(1, graph.edges.length);
             assert.equal(playerContractRelation, graph.edges[0].relation);
-        })
-    })
+        });
+    });
 
     describe('#removeEdges', () => {
         it('should remove a graph edge', () => {
-            const container = new JGFContainer(singleGraph = true);
-            const graph = container.graph;
+            const container = new JGFContainer(true);
+            const { graph } = container;
 
             const node1Id = 'lebron-james#2254';
             const node1Label = 'LeBron James';
@@ -178,13 +176,13 @@ describe('Graph', () => {
 
             graph.removeEdges(node1Id, node2Id, playerContractRelation);
             assert.equal(0, graph.edges.length, 'After removeEdges there should be zero edges');
-        })
-    })
+        });
+    });
 
     describe('#getEdges', () => {
         it('should lookup edges', () => {
-            const container = new JGFContainer(singleGraph = true);
-            const graph = container.graph;
+            const container = new JGFContainer(true);
+            const { graph } = container;
 
             const node1Id = 'lebron-james#2254';
             const node1Label = 'LeBron James';
@@ -201,23 +199,23 @@ describe('Graph', () => {
             const edges = graph.getEdges(node1Id, node2Id, playerContractRelation);
             assert(edges !== null);
             assert.equal(1, edges.length);
-        })
-    })
+        });
+    });
 
     describe('#graphDimensions', () => {
         it('should return zero dimensions for an empty graph', () => {
-            const container = new JGFContainer(singleGraph = true);
-            const graph = container.graph;
+            const container = new JGFContainer(true);
+            const { graph } = container;
 
             const dimensions = graph.graphDimensions;
             assert.equal(0, dimensions.nodes);
             assert.equal(0, dimensions.edges);
-        })
+        });
 
 
         it('should return valid dimensions for a non-empty graph', () => {
-            const container = new JGFContainer(singleGraph = true);
-            const graph = container.graph;
+            const container = new JGFContainer(true);
+            const { graph } = container;
 
             graph.addNode('node1', 'nodeTypeA');
             graph.addNode('node2', 'nodeTypeB');
@@ -226,7 +224,6 @@ describe('Graph', () => {
             const dimensions = graph.graphDimensions;
             assert.equal(2, dimensions.nodes);
             assert.equal(1, dimensions.edges);
-        })
-    })
-
+        });
+    });
 });