+++ test/004.testSave.js
@@ -1,24 +1,24 @@
 const { assert } = require('chai');
-const { JGFContainer } = require('../jgfContainer');
 const fsExtra = require('fs-extra');
 const path = require('path');
+const { JGFContainer } = require('../jgfContainer');
 
 /* eslint no-invalid-this: 0 */
 
 describe('ContainerSaveToFile', () => {
     describe('#saveToFile-onenode', () => {
-        beforeEach(async() => {
+        beforeEach(async () => {
             this.currentTest = {};
             this.currentTest.filename = './test/temp/test-onenode.json';
             await fsExtra.ensureDir(path.dirname(this.currentTest.filename));
             if (await fsExtra.exists(this.currentTest.filename)) {
                 await fsExtra.remove(this.currentTest.filename);
             }
-        })
+        });
 
         it('should save a graph that only has one node to a file', async () => {
-            const container = new JGFContainer(singleGraph = true);
-            const graph = container.graph;
+            const container = new JGFContainer(true);
+            const { graph } = container;
 
             const nodeId = 'lebron-james#2254';
             const nodeLabel = 'LeBron James';
@@ -36,8 +36,8 @@ describe('ContainerSaveToFile', () => {
             const container2 = new JGFContainer();
             container2.loadFromFile(this.currentTest.filename);
             assert.equal(1, container.graph.nodes.length);
-        })
-    })
+        });
+    });
 
     describe('#saveToFile-full', () => {
         beforeEach(async () => {
@@ -47,11 +47,11 @@ describe('ContainerSaveToFile', () => {
             if (await fsExtra.exists(this.currentTest.filename)) {
                 await fsExtra.remove(this.currentTest.filename);
             }
-        })
+        });
 
         it('should save a graph that has multiple nodes and edges to a file', async () => {
-            const container = new JGFContainer(singleGraph = true);
-            const graph = container.graph;
+            const container = new JGFContainer(true);
+            const { graph } = container;
 
             const node1Id = 'kyrie-irving#202681';
             const node1Label = 'Kyrie Irving';
@@ -71,7 +71,6 @@ describe('ContainerSaveToFile', () => {
             const fileContent = await fsExtra.readJson(this.currentTest.filename);
             assert.equal(2, fileContent.graph.nodes.length);
             assert.equal(1, fileContent.graph.edges.length);
-        })
-    })
-
+        });
+    });
 });